remove public prefix, add server only
