+++ evi-next-js-function-calling/.env.example
@@ -1,4 +1,11 @@
-NEXT_PUBLIC_HUME_API_KEY=""
-NEXT_PUBLIC_HUME_SECRET_KEY=""
-NEXT_PUBLIC_HUME_VOICE_WEATHER_CONFIG_ID=""
+# Used to fetch access token
+# do not add NEXT_PUBLIC_ prefix to avoid leaking your credentials
+HUME_API_KEY=""
+HUME_SECRET_KEY=""
+
+# API Key for the geocoding API
 NEXT_PUBLIC_GEOCODING_API_KEY=""
+
+# This is the config ID for the weather voice
+# This is not sensitive information and can be prefixed with NEXT_PUBLIC_
+NEXT_PUBLIC_HUME_VOICE_WEATHER_CONFIG_ID=""
