+++ evi-next-js-function-calling/app/globals.css
