+++ evi-next-js-function-calling/app/page.tsx
@@ -1,0 +1,13 @@
+// ./app/page.tsx
+import ClientComponent from '@/components/ClientComponent';
+import { getHumeAccessToken } from '@/utils/getHumeAccessToken';
+
+export default async function Page() {
+  const accessToken = await getHumeAccessToken();
+
+  if (!accessToken) {
+    throw new Error();
+  }
+
+  return <ClientComponent accessToken={accessToken} />;
+}
