+++ evi-next-js-function-calling/components/ClientComponent.tsx
