+++ evi-next-js-function-calling/package.json
@@ -13,7 +13,8 @@
     "@humeai/voice-react": "^0.1.7",
     "next": "14.2.3",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "server-only": "^0.0.1"
   },
   "devDependencies": {
     "@types/node": "^20.12.7",
