+++ evi-next-js-function-calling/utils/getHumeAccessToken.ts
@@ -1,0 +1,15 @@
+import "server-only";
+import { fetchAccessToken } from "@humeai/voice";
+
+export const getHumeAccessToken = async () => {
+  const accessToken = await fetchAccessToken({
+    apiKey: String(process.env.HUME_API_KEY),
+    secretKey: String(process.env.HUME_SECRET_KEY),
+  });
+
+  if (accessToken === "undefined") {
+    return null;
+  }
+
+  return accessToken ?? null;
+};
